Subscription receives changes based on origin
Add a subscription parameter (filter_origins) that controls from who
the subscription will receive changes based on the origin. If
filter_origins is empty or was not specified, forward changes from
any origin.

Origins can be obtained from the pg_replication_origin catalog on the
publisher node.

This feature is useful for bidirectional replication scenarios in which
we want to avoid sending data back to origin.